op-node: merge EngineResetDeriver with EngineController
